LayoutTests/http/tests/plugins/nounsupported-plugin.html

<html>
<body>
<script src= "../resources/testharness.js"></script>
<script src= "../resources/testharnessreport.js"></script>
<script>
function checkPlugins()
{
    return !!(navigator.plugins["CyberKit Test PlugIn"] || navigator.plugins["CyberKit built-in PDF"]);
}

test(() => {
    assert_true(checkPlugins());
}, "Test plugins");
</script>
</body>
</html>
